+++ demo/pages/_app.tsx
@@ -33,6 +33,9 @@ const REFERRAL_FEE = process.env.NEXT_PUBLIC_REFERRAL_FEE
   : undefined
 const REFERRAL_FEE_RECIPIENT =
   process.env.NEXT_PUBLIC_REFERRAL_FEE_RECIPIENT || undefined
+const NORMALIZE_ROYALTIES = process.env.NEXT_PUBLIC_NORMALIZE_ROYALTIES
+  ? process.env.NEXT_PUBLIC_NORMALIZE_ROYALTIES === 'true'
+  : false
 const ALCHEMY_KEY = process.env.NEXT_PUBLIC_ALCHEMY_KEY || undefined
 
 const envChain = allChains.find((chain) => chain.id === +CHAIN_ID)
@@ -81,6 +84,7 @@ const AppWrapper = ({ children }) => {
         referralFee: REFERRAL_FEE,
         referralFeeRecipient: REFERRAL_FEE_RECIPIENT,
         source: SOURCE,
+        normalizeRoyalties: NORMALIZE_ROYALTIES,
       }}
       theme={theme}
     >