+++ demo/pages/modal/acceptBid.tsx
@@ -10,6 +10,9 @@ const DEFAULT_COLLECTION_ID =
   process.env.NEXT_PUBLIC_DEFAULT_COLLECTION_ID ||
   '0xe14fa5fba1b55946f2fa78ea3bd20b952fa5f34e'
 const DEFAULT_TOKEN_ID = process.env.NEXT_PUBLIC_DEFAULT_TOKEN_ID || '39'
+const NORMALIZE_ROYALTIES = process.env.NEXT_PUBLIC_NORMALIZE_ROYALTIES
+  ? process.env.NEXT_PUBLIC_NORMALIZE_ROYALTIES === 'true'
+  : false
 
 const AcceptBidPage: NextPage = () => {
   const router = useRouter()
@@ -18,6 +21,8 @@ const AcceptBidPage: NextPage = () => {
   const deeplinkOpenState = useState(true)
   const hasDeeplink = router.query.deeplink !== undefined
   const [bidId, setBidId] = useState('')
+  const [normalizeRoyalties, setNormalizeRoyalties] =
+    useState(NORMALIZE_ROYALTIES)
 
   useEffect(() => {
     const prefilledBidId = router.query.bidId
@@ -69,6 +74,16 @@ const AcceptBidPage: NextPage = () => {
         />
       </div>
       <DeeplinkCheckbox />
+      <div>
+        <label>Normalize Royalties: </label>
+        <input
+          type="checkbox"
+          checked={normalizeRoyalties}
+          onChange={(e) => {
+            setNormalizeRoyalties(e.target.checked)
+          }}
+        />
+      </div>
 
       <AcceptBidModal
         trigger={
@@ -92,6 +107,7 @@ const AcceptBidPage: NextPage = () => {
         tokenId={tokenId}
         openState={hasDeeplink ? deeplinkOpenState : undefined}
         bidId={bidId}
+        normalizeRoyalties={normalizeRoyalties}
         onBidAccepted={(data) => {
           console.log('Bid Accepted', data)
         }}