+++ packages/ui/src/hooks/useCollections.ts
@@ -6,16 +6,24 @@ type CollectionResponse =
   paths['/collections/v5']['get']['responses']['200']['schema']
 
 export default function (
-  query?: paths['/collections/v5']['get']['parameters']['query'] | false,
+  options?: paths['/collections/v5']['get']['parameters']['query'] | false,
   swrOptions: SWRConfiguration = {}
 ) {
   const client = useReservoirClient()
 
   const path = new URL(`${client?.apiBase}/collections/v5`)
-  setParams(path, query || {})
+  const query = options || {}
+  if (
+    query.normalizeRoyalties === undefined &&
+    client?.normalizeRoyalties !== undefined
+  ) {
+    query.normalizeRoyalties = client.normalizeRoyalties
+  }
+
+  setParams(path, query)
 
   const { data, mutate, error, isValidating } = useSWR<CollectionResponse>(
-    query ? [path.href, client?.apiKey, client?.version] : null,
+    options ? [path.href, client?.apiKey, client?.version] : null,
     null,
     {
       revalidateOnMount: true,