+++ packages/ui/src/modal/buy/BuyModalRenderer.tsx
@@ -66,6 +66,7 @@ type Props = {
   collectionId?: string
   referrerFeeBps?: number | null
   referrer?: string | null
+  normalizeRoyalties?: boolean
   children: (props: ChildrenProps) => ReactNode
 }
 
@@ -75,6 +76,7 @@ export const BuyModalRenderer: FC<Props> = ({
   collectionId,
   referrer,
   referrerFeeBps,
+  normalizeRoyalties,
   children,
 }) => {
   const { data: signer } = useSigner()
@@ -94,6 +96,7 @@ export const BuyModalRenderer: FC<Props> = ({
   const { data: tokens } = useTokens(
     open && {
       tokens: [`${contract}:${tokenId}`],
+      normalizeRoyalties,
     },
     {
       revalidateFirstPage: true,
@@ -102,6 +105,7 @@ export const BuyModalRenderer: FC<Props> = ({
   const { data: collections } = useCollections(
     open && {
       id: collectionId,
+      normalizeRoyalties,
     }
   )
   const collection = collections && collections[0] ? collections[0] : undefined
@@ -148,6 +152,10 @@ export const BuyModalRenderer: FC<Props> = ({
       options.feesOnTop = [`${referrer}:${referrerFeeBps}`]
     } else if (referrer === null && referrerFeeBps === null) {
       delete options.feesOnTop
+    }
+
+    if (normalizeRoyalties !== undefined) {
+      options.normalizeRoyalties = normalizeRoyalties
     }
 
     setBuyStep(BuyStep.Approving)
@@ -215,7 +223,12 @@ export const BuyModalRenderer: FC<Props> = ({
         if (error && error?.message.includes('ETH balance')) {
           setHasEnoughCurrency(false)
         } else {
-          const transactionError = new Error(error?.message || '', {
+          const errorType = (error as any)?.type
+          let message = 'Oops, something went wrong. Please try again.'
+          if (errorType && errorType === 'price mismatch') {
+            message = error.message
+          }
+          const transactionError = new Error(message, {
             cause: error,
           })
           setTransactionError(transactionError)
@@ -230,6 +243,7 @@ export const BuyModalRenderer: FC<Props> = ({
     collectionId,
     referrer,
     referrerFeeBps,
+    normalizeRoyalties,
     client,
     signer,
     currency,