+++ packages/ui/src/modal/list/ListModalRenderer.tsx
@@ -85,6 +85,7 @@ type Props = {
   tokenId?: string
   collectionId?: string
   currencies?: Currency[]
+  normalizeRoyalties?: boolean
   children: (props: ChildrenProps) => ReactNode
 }
 
@@ -138,6 +139,7 @@ export const ListModalRenderer: FC<Props> = ({
   tokenId,
   collectionId,
   currencies,
+  normalizeRoyalties,
   children,
 }) => {
   const { data: signer } = useSigner()
@@ -172,13 +174,14 @@ export const ListModalRenderer: FC<Props> = ({
   )
 
   const [expirationOption, setExpirationOption] = useState<ExpirationOption>(
-    expirationOptions[0]
+    expirationOptions[5]
   )
 
   const { data: tokens } = useTokens(
     open && {
       tokens: [`${contract}:${tokenId}`],
       includeAttributes: true,
+      normalizeRoyalties,
     },
     {
       revalidateFirstPage: true,
@@ -187,6 +190,7 @@ export const ListModalRenderer: FC<Props> = ({
   const { data: collections } = useCollections(
     open && {
       id: collectionId,
+      normalizeRoyalties,
     }
   )
 
@@ -376,7 +380,7 @@ export const ListModalRenderer: FC<Props> = ({
       }
       setLoadedInitalPrice(false)
       setStepData(null)
-      setExpirationOption(expirationOptions[0])
+      setExpirationOption(expirationOptions[5])
       setSyncProfit(true)
     }
     setCurrency(currencies && currencies[0] ? currencies[0] : defaultCurrency)